requirements.txt
django
gunicorn
python-dotenv
whitenoise
git+https://${GITHUB_TOKEN_SOFT_UI}@github.com/app-generator/priv-django-admin-soft-pro@v1.0.11#egg=django-admin-soft-pro

psycopg2-binary
docx-mailmerge
# mysqlclient
